feat: add switching the view of catalog
