+++ src/components/Card/card.css
@@ -12,6 +12,11 @@
   flex-direction: column;
 }
 
+.small {
+  width: 14.6%;
+  padding-top: 12%;
+}
+
 .card__wrap {
   width: 100%;
   height: 100%;
@@ -24,6 +29,11 @@
   width: 100%;
   height: 83%;
   cursor: pointer;
+  opacity: 1;
+}
+
+.card__text-none {
+  opacity: 0;
 }
 
 .card__title {
@@ -74,4 +84,8 @@
   padding-right: 2%;
   font-size: 80%;
   cursor: pointer;
+}
+
+.button__small {
+  font-size: 65%;
 }
