+++ src/components/Catalog/catalog.css
@@ -68,6 +68,11 @@
   user-select: none;
 }
 
+.active {
+  box-shadow: 0 0 3px 1px #c45f07;
+  border: 1px solid rgb(235, 68, 68);
+}
+
 .products__wrap {
   padding: 0.5%;
   display: flex;
